requirements.txt
Django==5.1.2
numpy==1.26.4
pandas==2.2.0
matplotlib==3.10.8
seaborn==0.13.2
pillow==10.2.0
tensorflow-cpu==2.15.0
scikit-learn==1.5.2
opencv-python-headless==4.10.0.84
requests==2.32.5
h5py==3.14.0
gunicorn==21.2.0
whitenoise==6.6.0
